backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.44
psycopg[binary]==3.2.11
alembic==1.14.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic>=2.10.0
pydantic-settings>=2.3.0

# Real-time
python-socketio==5.10.0
python-engineio==4.8.0

# External Services
requests==2.32.0
python-dotenv==1.0.1
cloudinary==1.41.0
httpx==0.27.0

# Utilities
pytz==2023.3
python-dateutil==2.8.2
email-validator==2.1.0

# Development & Testing
pytest>=8.2
pytest-asyncio==0.24.0
pytest-cov==5.0.0
black==24.10.0
flake8==7.1.1

# Monitoring
sentry-sdk==2.0.0

# Task Queue (optional for MVP)
redis==5.1.1
celery==5.4.0